# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервер для управления пользователями",
        "title": "My Project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает нового пользователя с переданными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/pfo-vector_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса или валидация не пройдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким telegram_id уже существует",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные пользователей по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение всех пользователей",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователей с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/pfo-vector_internal_model.UserListResponse"
                        }
                    },
                    "404": {
                        "description": "Ошибка получения списка пользователей",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/all/Name": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные пользователей по Full_name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение всех пользователей",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователей с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/pfo-vector_internal_model.UserListResponse"
                        }
                    },
                    "404": {
                        "description": "Ошибка получения списка пользователей",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/all/Rating": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные пользователей по Rating",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение всех пользователей",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователей с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/pfo-vector_internal_model.UserListResponse"
                        }
                    },
                    "404": {
                        "description": "Ошибка получения списка пользователей",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/import-users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает Excel-файл и импортирует пользователей",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Импорт пользователей из Excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel файл (.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Импорт завершён",
                        "schema": {
                            "$ref": "#/definitions/pfo-vector_internal_model.ImportResult"
                        }
                    },
                    "400": {
                        "description": "Некорректный файл или данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Конфликт уникальных данных",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.CreateUserRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "direction_vector": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "description": "Если nil -\u003e БД поставит 0",
                    "type": "integer"
                },
                "study_group": {
                    "type": "string"
                },
                "telegram": {
                    "description": "Обязательно (NOT NULL в БД)",
                    "type": "string"
                },
                "visited_events_count": {
                    "description": "Если nil -\u003e БД поставит 0",
                    "type": "integer"
                }
            }
        },
...
        "pfo-vector_internal_model.ImportResult": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pfo-vector_internal_model.RowError"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "pfo-vector_internal_model.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "pfo-vector_internal_model.RowError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        },
        "pfo-vector_internal_model.UserListResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/pfo-vector_internal_model.Pagination"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pfo-vector_internal_model.UserResponse"
                    }
                }
            }
        },
        "pfo-vector_internal_model.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "direction_vector": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "join_date": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "study_group": {
                    "type": "string"
                },
                "telegram": {
                    "type": "string"
                },
                "visited_events_count": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}